requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
PyJWT
httpx
sqlalchemy
passlib
pymssql
bcrypt